# swagger.yaml
swagger: '2.0'
info:
  version: '1.0.0'
  title: Workshop App
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
  /users:
    get:
      summary: Userの一覧を得る
      operationId: getUsers
      description: |
        get list of user 
      responses:
        200:
          description: OK
          schema:
            type: "array"
            items:
              $ref: '#/definitions/User'
        400:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
    post:
      summary: Userを一件追加する
      operationId: addUser
      description: |
        add a user
      parameters:
        - name: body
          in: body
          description: user object
          required: true
          schema:
             $ref: '#/definitions/User'
      responses:
        201:
          description: OK
        400:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'

  /users/{user_id}/:
    get:
      summary: idを使って特定のUserを得る
      operationId: getUserById
      description: |
        get a user by id
      parameters:
        - name: user_id
          in: path
          required: true
          type: integer
          description: ID of user
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/User'
        400:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
    put:
      summary: idを使って特定のUserを変更する
      operationId: updateUser
      description: |
        update a user
      parameters:
        - name: user_id
          in: path
          required: true
          type: integer
          description: ID of user
        - name: body
          in: body
          description: user object
          required: true
          schema:
             $ref: '#/definitions/User'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/User'
        400:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
definitions:
  User:
    type: object
    required:
      - id
      - userId
      - password
    properties:
      id:
        type: integer
      userId:
        type: string
      password:
        type: string
  ErrorResponse:
    type: object
    required:
      - code
      - name
      - message
    properties:
      code:
        type: integer
      name:
        type: string
      message:
        type: string